aero-python/pyproject.toml
[tool.poetry]
name = "aerorule"
version = "0.1.0"
description = "AeroRule evaluation via CEL"
authors = ["Author <author@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
cel-python = "*"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
